# web/server/docs/openapi-v1.yaml
openapi: 3.0.0
info:
    title: ApexCare API
    description: API for ApexCare Solutions, a maintainance service solutions provider connecting users with skilled technicians ensuring that users' equiptment are well maintained.
    version: 1.0.0
    contact:
        name: ApexCare Support
        url: https://apexcare.co.za/support

servers:
    - url: https://apexcare.co.za/api/v1

security:
    - bearerAuth: []

components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT

    schemas:
        Job:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                type:
                    type: string
                    example: "Electrical Repair"
                service:
                    type: string
                    example: "Wiring Installation"
                location:
                    type: string
                    example: "123 Main Street"
                priority:
                    type: string
                    example: "High"
                status:
                    type: string
                    example: "PENDING"

        Technician:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                firstName:
                    type: string
                    example: "John"
                lastName:
                    type: string
                    example: "Doe"
                phone:
                    type: string
                    example: "+123456789"
                preferredContact:
                    type: string
                    enum: [EMAIL, PHONE, SMS, NONE]
                    example: "SMS"

        CreateJobRequest:
            type: object
            properties:
                job:
                    $ref: "#/components/schemas/Job"
                technician:
                    $ref: "#/components/schemas/Technician"

paths:
    /jobs:
        post:
            summary: Create a new job
            description: Creates a new job and notifies the assigned technician.
            operationId: createJob
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateJobRequest"
            responses:
                "201":
                    description: Job successfully created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Job"
                "400":
                    description: Bad request

    /jobs/accept/{id}:
        post:
            summary: Accept a job
            description: Accepts a job and notifies the user that the technician is on their way.
            operationId: acceptJob
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      example: 1
            responses:
                "201":
                    description: Job successfully accepted
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Job"
                "400":
                    description: Bad request
                "404":
                    description: Job not found

    /jobs/decline/{id}:
        post:
            summary: Decline a job
            description: Declines a job and updates its status.
            operationId: declineJob
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      example: 1
            responses:
                "201":
                    description: Job successfully declined
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Job"
                "400":
                    description: Bad request
                "404":
                    description: Job not found